refactor(client): colocate single-route components into route folders
Move shared components used by exactly one route family into that route's
*.-components/ folder (found by /check-components, #321):

- WeeklyEntryEditor       -> routines.$id.edit.-components
- DailyDetailsPanel       -> routines.$id.-components
- TodosChecklist          -> tasks.$id.-components
- ResourcesTable          -> tasks.$id.-components
- ResourceInteractionsLog -> resources.$id.-components (new folder)
- ResourceModulesAdmin    -> resources.$id.-components (new folder)

Shared sibling helpers stay in components/ and are imported via the @/
alias. Barrel re-exports, consumer imports, and co-located stories were
updated accordingly. routines.$id.index.tsx gains an import/max-dependencies
disable, matching the sibling resources route pages.

Two checklist items are intentionally left out as no longer valid:
DailiesActiveListView is now consumed by two route families (dashboard +
routines tracker), and ResourceLinksPicker is shared through the formFields
barrel (colocating it would trip import/max-dependencies in -TopicForm).

Refs #321
