+++ packages/client/src/routes/resources.$id.-components/-ResourceInteractionsLog.tsx
