+++ packages/client/src/routes/resources.$id.-components/-ResourceModulesAdmin.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 
 import { expect, within } from "storybook/test";
 
-import { ResourceModulesAdmin } from "./ResourceModulesAdmin";
+import { ResourceModulesAdmin } from "./-ResourceModulesAdmin";
 
 import {
   makeModule,
