+++ packages/client/src/routes/routines.$id.edit.-components/-WeeklyEntryEditor.tsx
