+++ packages/client/src/routes/tasks.$id.-components/-ResourcesTable.tsx
@@ -5,15 +5,14 @@ import { useMemo, useState } from "react";
 
 import { PlusIcon, SearchIcon } from "lucide-react";
 
+import { Input } from "@/components/input";
 import {
   COLUMN_COUNT,
   EditingRow,
   inheritedLevel,
   linkedResourceLabel,
   TaskResourceRow,
-} from "./taskResourceTable";
-
-import { Input } from "@/components/input";
+} from "@/components/tasks/taskResourceTable";
 import { Button } from "@/components/ui/button";
 import { DataTable } from "@/components/ui/data-table";
 import {
