+++ packages/client/src/routes/tasks.$id.-components/-TodosChecklist.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 
 import { expect, within } from "storybook/test";
 
-import { TodosChecklist } from "./TodosChecklist";
+import { TodosChecklist } from "./-TodosChecklist";
 
 import { makeTask } from "@/test-utils/boxFixtures";
 import { QueryStub } from "@/test-utils/QueryStub";
